move things around, fix tests
