+++ packages/consent-manager/components/banner.js
