+++ packages/consent-manager/components/banner.test.js
