+++ packages/consent-manager/index.test.js
@@ -1,5 +1,4 @@
-import { fireEvent, render, screen } from '@testing-library/react'
-import ConsentManager from './'
+import { fireEvent, render, screen, act } from '@testing-library/react'
 
 const defaultProps = {
   version: 0,
@@ -22,8 +21,7 @@ const defaultProps = {
       category: 'Example Category',
       description:
         'A short description of what the service is and how your company uses the data.',
-      body:
-        '// a chunk of javascript to add to the page if permission is granted\n// this is optional',
+      body: '',
       url: 'http://www.an-optional-url-for-a-script-to-add-to-the-page.com',
     },
     {
@@ -52,13 +50,14 @@ const defaultProps = {
 }
 
 test('shows the banner if the forceShow prop is true', () => {
+  const { default: ConsentManager } = require('./')
   render(<ConsentManager {...defaultProps} forceShow={true} />)
   expect(screen.getByTestId('consent-banner')).toBeInTheDocument()
 })
 
 test('sets existing preferences and does not show the banner if preferences are already set', async () => {
   await runWithMockedImport(
-    './partials/cookies',
+    './util/cookies',
     {
       loadPreferences: () => {
         return { loadAll: false, segment: false, version: 0 }
@@ -74,7 +73,7 @@ test('sets existing preferences and does not show the banner if preferences are 
 
 test('shows the banner if preferences are set but version has increased', async () => {
   await runWithMockedImport(
-    './partials/cookies',
+    './util/cookies',
     {
       loadPreferences: () => {
         return { loadAll: false, segment: false, version: 1 }
@@ -92,7 +91,7 @@ test('shows the banner if preferences are set but version has increased', async 
 
 test('if the manage preferences button is clicked, opens the dialog and makes body un-scrollable', async () => {
   await runWithMockedImport(
-    './partials/dialog',
+    './components/dialog',
     () => {
       return <p data-testid="mocked-dialog">test</p>
     },
@@ -107,13 +106,13 @@ test('if the manage preferences button is clicked, opens the dialog and makes bo
 
 test('opens the dialog when the open function is called', async () => {
   await runWithMockedImport(
-    './partials/dialog',
+    './components/dialog',
     () => {
       return <p data-testid="mocked-dialog">test</p>
     },
     async (MockedConsentManager, open) => {
       render(<MockedConsentManager {...defaultProps} />)
-      open()
+      act(() => open())
       expect(screen.getByTestId('mocked-dialog')).toBeInTheDocument()
       expect(document.body.className).toBe('g-noscroll')
     }
@@ -123,7 +122,7 @@ test('opens the dialog when the open function is called', async () => {
 test('if accept all button is clicked, any open dialogs are closed and all analytics are loaded', async () => {
   let prefs
   await runWithMockedImport(
-    './partials/cookies',
+    './util/cookies',
     {
       loadPreferences: () => {},
       savePreferences: (preferences) => {
@@ -138,12 +137,43 @@ test('if accept all button is clicked, any open dialogs are closed and all analy
   expect(prefs.loadAll).toBe(true)
 })
 
+test('loads segment and additional services if loadAll is passed', async () => {
+  await runWithMockedImport(
+    './util/cookies',
+    {
+      loadPreferences: () => {
+        return { loadAll: true, segment: { foo: 'bar' } }
+      },
+      savePreferences: () => {},
+    },
+    (MockedConsentManager) => {
+      render(
+        <MockedConsentManager {...defaultProps} forceShow={false} version={2} />
+      )
+      const html = document.body.innerHTML
+      // script was injected
+      expect(html).toMatch(
+        /<script type="text\/javascript" src="https:\/\/cdn.segment.com\/analytics\.js/
+      )
+      // all services loaded, as well as custom segment
+      expect(html).toMatch(
+        /analytics\.load\("iyi06c432UL7SB1r3fQReec4bNwFyzkW", {"integrations":{"All":true,"Segment\.io":true,"foo":"bar"}}\);/
+      )
+
+      // custom service loaded
+      expect(html).toMatch(
+        /src="http:\/\/www.an-optional-url-for-a-script-to-add-to-the-page\.com"/
+      )
+    }
+  )
+})
+
 // Given an internal module name and mock implementation, mocks the given module and returns a version
 // of the component with the module mocked.
 async function runWithMockedImport(module, mock, cb) {
-  jest.resetModules()
-  jest.mock(module, () => mock)
-  const { default: ConsentManager, open } = await import('./')
-  cb(ConsentManager, open)
-  jest.resetAllMocks()
+  jest.isolateModules(() => {
+    jest.doMock(module, () => mock)
+    const { default: ConsentManager, open } = require('./')
+    cb(ConsentManager, open)
+  })
 }
