+++ packages/consent-manager/util/integrations.js
